refactor: base opeators in in-place ones
Operators should be implemented based on their in-place variants, not the other way around. In that way in-place operations won't allocate any new objects.

See #916 for more details.

This PR allows us to reenable the swiftlint rule shorthand_operator. It also refactors other operators to use their in-place variant to reduce code duplication.